Support skipping some tokens during parsing
This makes it easy to use sagnificant whitespace.  The parser will skip over
a token (in our case a newline) when it doesn't match the terminal on the
top of the stack, or is not found in the parsing table (when a non-terminal
is on the top of the stack).

src/parsing.m:
    As above.

src/parse.m:
    Skip newlines.