WIP: testing and fixing lookahead.
I had been working on this for a while.  I think I know how to finish it but
it's just going to be easier to alter the grammar.  The language will be a
little clumsy as a result, until we replace the parsing code completely.
But I'll commit this to a separate branch in case it becomes useful.